address comments on 4th reviews
