+++ internal/manager/featuregates/feature_gates_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/go-logr/zapr"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 )
 
@@ -27,7 +28,7 @@ func TestFeatureGates(t *testing.T) {
 	t.Log("Verifying feature gates setup will return error when settings has conflicts")
 	featureGates = map[string]bool{KongCustomEntity: true, FillIDsFeature: false}
 	_, err = New(setupLog, featureGates)
-	assert.Error(t, err)
+	require.Error(t, err)
 	assert.Contains(t, err.Error(), fmt.Sprintf("%s is required if %s is enabled", FillIDsFeature, KongCustomEntity))
 
 	t.Log("Configuring several invalid feature gates options")
@@ -35,6 +36,6 @@ func TestFeatureGates(t *testing.T) {
 
 	t.Log("Verifying feature gates setup results when invalid feature gates options are present")
 	_, err = New(setupLog, featureGates)
-	assert.Error(t, err)
+	require.Error(t, err)
 	assert.Contains(t, err.Error(), "invalidGateway is not a valid feature")
 }
